make image and point cloud viewers. Need to fix & test
