+++ app/src/js/components/image_viewer.tsx
@@ -1,7 +1,15 @@
 import { withStyles } from '@material-ui/styles'
 import React from 'react'
+import { changeViewerConfig } from '../action/common'
+import Session from '../common/session'
+import * as types from '../common/types'
 import { ImageViewerConfigType } from '../functional/types'
 import { viewerStyles } from '../styles/viewer'
+import {
+  MAX_SCALE,
+  MIN_SCALE,
+  SCROLL_ZOOM_RATIO
+} from '../view_config/image'
 import { DrawableViewer, ViewerProps } from './drawable_viewer'
 import ImageCanvas from './image_canvas'
 import Label2dCanvas from './label2d_canvas'
@@ -16,24 +24,6 @@ class ImageViewer extends DrawableViewer {
    */
   constructor (props: ViewerProps) {
     super(props)
-  }
-
-  /**
-   * Run when component mounts
-   */
-  public componentDidMount () {
-    super.componentDidMount()
-    document.addEventListener('keydown', this._keyDownHandler)
-    document.addEventListener('keyup', this._keyUpHandler)
-  }
-
-  /**
-   * Run when component unmounts
-   */
-  public componentWillUnmount () {
-    super.componentWillUnmount()
-    document.removeEventListener('keydown', this._keyDownHandler)
-    document.removeEventListener('keyup', this._keyUpHandler)
   }
 
   /**
@@ -52,14 +42,14 @@ class ImageViewer extends DrawableViewer {
     if (this._viewerConfig) {
       views.push(
         <ImageCanvas
-          key={`imageView${this.props.id}`}
+          key={`imageCanvas${this.props.id}`}
           display={this._container}
           id={this.props.id}
         />
       )
       views.push(
         <Label2dCanvas
-          key={`imageView${this.props.id}`}
+          key={`label2dCanvas${this.props.id}`}
           display={this._container}
           id={this.props.id}
         />
@@ -70,53 +60,45 @@ class ImageViewer extends DrawableViewer {
   }
 
   /**
-   * Handle mouse down
-   * @param e
-   */
-  protected onMouseDown (e: React.MouseEvent) {
-    if (e.button === 2) {
-      e.preventDefault()
-    }
-    this._viewerConfigUpdater.onMouseDown(e.clientX, e.clientY, e.button)
-  }
-
-  /**
-   * Handle mouse up
-   * @param e
-   */
-  protected onMouseUp (_e: React.MouseEvent) {
-    this._viewerConfigUpdater.onMouseUp()
-  }
-
-  /**
    * Handle mouse move
    * @param e
    */
   protected onMouseMove (e: React.MouseEvent) {
-    this._viewerConfigUpdater.onMouseMove(e.clientX, e.clientY)
+    const oldX = this._mX
+    const oldY = this._mY
+    super.onMouseMove(e)
+    if (this._mouseDown && this._container && this._viewerConfig) {
+      if (this.isKeyDown(types.Key.META) ||
+          this.isKeyDown(types.Key.CONTROL)) {
+        const dx = this._mX - oldX
+        const dy = this._mY - oldY
+        const displayLeft = this._container.scrollLeft - dx
+        const displayTop = this._container.scrollTop - dy
+        const newConfig = {
+          ...this._viewerConfig,
+          displayLeft,
+          displayTop
+        }
+        Session.dispatch(changeViewerConfig(
+          this._viewerId, newConfig
+        ))
+      }
+    }
   }
 
   /**
    * Handle double click
    * @param e
    */
-  protected onDoubleClick (e: React.MouseEvent) {
-    this._viewerConfigUpdater.onDoubleClick(e.clientX, e.clientY)
+  protected onDoubleClick () {
+    return
   }
 
   /**
    * Handle mouse leave
    * @param e
    */
-  protected onMouseEnter (_e: React.MouseEvent) {
-    Session.activeViewerId = this.props.id
-  }
-
-  /**
-   * Handle mouse leave
-   * @param e
-   */
-  protected onMouseLeave (_e: React.MouseEvent) {
+  protected onMouseLeave () {
     return
   }
 
@@ -126,29 +108,34 @@ class ImageViewer extends DrawableViewer {
    */
   protected onWheel (e: React.WheelEvent) {
     e.preventDefault()
-    this._viewerConfigUpdater.onWheel(e.deltaY)
-  }
-
-  /**
-   * Handle key down
-   * @param e
-   */
-  protected onKeyUp (e: KeyboardEvent) {
-    if (Session.activeViewerId === this.props.id) {
-      this._viewerConfigUpdater.onKeyUp(e.key)
+    if (this._viewerConfig) {
+      if (this.isKeyDown(types.Key.META) ||
+          this.isKeyDown(types.Key.CONTROL)) {
+        let zoomRatio = SCROLL_ZOOM_RATIO
+        if (e.deltaY < 0) {
+          zoomRatio = 1. / zoomRatio
+        }
+        const config = this._viewerConfig as ImageViewerConfigType
+        const newScale = config.viewScale * zoomRatio
+        const newConfig = { ...config }
+        if (newScale >= MIN_SCALE && newScale <= MAX_SCALE) {
+          newConfig.viewScale = newScale
+        }
+        if (this._container) {
+          const displayLeft = zoomRatio * (this._mX + config.displayLeft) -
+            this._mX
+          const displayTop = zoomRatio * (this._mY + config.displayTop) -
+            this._mY
+          newConfig.displayLeft = displayLeft
+          newConfig.displayTop = displayTop
+        }
+        Session.dispatch(changeViewerConfig(
+          this._viewerId,
+          newConfig
+        ))
+      }
     }
   }
-
-  /**
-   * Handle key down
-   * @param e
-   */
-  protected onKeyDown (e: KeyboardEvent) {
-    if (Session.activeViewerId === this.props.id) {
-      this._viewerConfigUpdater.onKeyDown(e.key)
-    }
-  }
-
 }
 
 export default withStyles(viewerStyles)(ImageViewer)
